requirements.txt
# Core dependencies
aiohttp>=3.9.0
feedparser>=6.0.0
pyyaml>=6.0
jinja2>=3.1.0
anthropic>=0.18.0
